+++ demos/world/src/main/java/is/codion/demos/world/domain/WorldImpl.java
@@ -216,7 +216,7 @@ public final class WorldImpl extends DomainModel {
 										Country.FLAG.define()
 														.column()
 														.caption("Flag")
-														.lazy(true),
+														.selected(false),
 										// end::lazy[]
 										Country.CODE_2.define()
 														.column()
@@ -322,7 +322,7 @@ public final class WorldImpl extends DomainModel {
 										Lookup.COUNTRY_FLAG.define()
 														.column()
 														.caption("Flag")
-														.lazy(true),
+														.selected(false),
 										Lookup.COUNTRY_CODE2.define()
 														.column()
 														.caption("Code2"),